fixed the slime chunk algorithm

+++ src/main/java/amidst/mojangapi/world/oracle/SlimeChunkOracle.java
@@ -13,11 +13,15 @@ public class SlimeChunkOracle {
 	}
 
 	public boolean isSlimeChunk(long chunkX, long chunkY) {
-		Random random = new Random(getSeed(chunkX, chunkY));
+		Random random = new Random(getSeed((int) chunkX, (int) chunkY));
 		return isSlimeChunk(random);
 	}
 
-	private long getSeed(long chunkX, long chunkY) {
+	/**
+	 * Make sure this uses integers for the parameters chunkX and chunkY and
+	 * long for the seed. This ensure the overflow works as in Minecraft.
+	 */
+	private long getSeed(int chunkX, int chunkY) {
 		return seed + chunkX * chunkX * 0x4c1906 + chunkX * 0x5ac0db + chunkY
 				* chunkY * 0x4307a7L + chunkY * 0x5f24f ^ 0x3ad8025f;
 	}
